pyproject.toml
[project]
name = "nomos-block-explorer"
version = "0.1.0"
requires-python = ">=3.14,<3.15"
dependencies = [
    "fastapi~=0.118.0",
    "httpx>=0.28.1",
    "pydantic-settings>=2.11.0",
    "python-on-whales~=0.78.0",
    "requests~=2.32.5",
    "rusty-results~=1.1.1",
    "sqlmodel~=0.0.25",
    "uvicorn~=0.37.0",
]

[tool.pyright]
include = ["src"]

[tool.black]
line-length = 120
target-version = ["py313"]
skip-string-normalization = false

[tool.isort]
profile = "black"
combine_as_imports = true
src_paths = ["src"]
skip_gitignore = true
ensure_newline_before_comments = true
